requirements.txt
flask==1.1.1
pymongo==3.10.1
pyyaml==5.3
